+++ packages/odyssey-react-mui/src/Select.tsx
@@ -62,10 +62,9 @@ export const SelectTestSelectors = {
           selector: {
             method: "ByRole",
             options: {
-              name: "${label}",
+              label: "name",
             },
             role: "option",
-            templateVariableNames: ["label"],
           },
         },
       },
@@ -80,10 +79,9 @@ export const SelectTestSelectors = {
   selector: {
     method: "ByRole",
     options: {
-      name: "${label}",
+      label: "name",
     },
     role: "combobox",
-    templateVariableNames: ["label"],
   },
 } as const satisfies FeatureTestSelector;
